cli: docker support
