+++ cli/pkg/cmd/destroy.go
@@ -2,12 +2,9 @@ package cmd
 
 import (
 	"context"
-
-	"github.com/defang-io/defang/cli/pkg/aws/ecs/pulumi"
-	"github.com/pulumi/pulumi-aws/sdk/v5/go/aws"
 )
 
-func Destroy(ctx context.Context, color Color, region aws.Region) error {
-	awsecs := pulumi.New(stack, region)
-	return awsecs.Destroy(ctx, pulumi.Color(color))
+func Destroy(ctx context.Context, color Color, region Region) error {
+	driver := createDriver(color, region)
+	return driver.Destroy(ctx)
 }
