+++ cli/pkg/cmd/logs.go
@@ -4,10 +4,9 @@ import (
 	"context"
 
 	"github.com/defang-io/defang/cli/pkg/aws/ecs/pulumi"
-	"github.com/pulumi/pulumi-aws/sdk/v5/go/aws"
 )
 
-func Logs(ctx context.Context, arn pulumi.TaskArn, region aws.Region) error {
-	awsecs := pulumi.New(stack, region)
-	return awsecs.Tail(ctx, arn)
+func Logs(ctx context.Context, arn pulumi.TaskArn, region Region) error {
+	driver := createDriver("auto", region)
+	return driver.Tail(ctx, arn)
 }
