+++ cli/pkg/docker/common.go
@@ -1,0 +1,27 @@
+package docker
+
+import (
+	"github.com/defang-io/defang/cli/pkg/types"
+	"github.com/docker/docker/client"
+)
+
+type ContainerID = types.TaskID
+
+type Docker struct {
+	*client.Client
+
+	image string
+}
+
+func New() *Docker {
+	cli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
+	if err != nil {
+		panic(err)
+	}
+
+	return &Docker{
+		Client: cli,
+	}
+}
+
+var _ types.Driver = (*Docker)(nil)
