+++ cli/pkg/docker/setup.go
@@ -1,0 +1,10 @@
+package docker
+
+import (
+	"context"
+)
+
+func (d *Docker) SetUp(ctx context.Context, image string) error {
+	d.image = image
+	return nil
+}
